Implement breadcrumbs and grouptree sync.
When clicking on a camembert part, this will leave a breadcrumb trail
at the top which can be clicked to go back to any of the parents.
The group tree will also highlight and open automatically the selected
group.

The grouptree doesn't have the camembert move when opening a
group, because it's unclear whether it's desirable at this point, though
probably it is. Ideally, clicking on the name would move to that
group, and clicking on the arrow would fold/unfold.