Implement the original camembert view
This is a shader and it implements a simple camembert view with
a basic animation when clicking and labels.
The view is just a circle diagram viewed from above, which might
not be as sophisticated as I'd like but it's plenty usable.
The labels are mostly fine, though the algorithm isn't great (neither
in terms of runtime, simplicity or result). They'll serve for now.

Also remove useless links which got in the repo by mistake